Add NewEventForm functionality
